중간고사 수정 적용(1차,경민)

+++ dongyangSD-master/app/src/main/java/com/example/campmate/ui/mypage/ReservationListScreen.kt
@@ -44,7 +44,7 @@ import com.example.campmate.data.model.Reservation
 @Composable
 fun ReservationListScreen(
     // ✅ [수정] campsiteName도 함께 전달하도록 변경
-    onNavigateToWriteReview: (campsiteId: Int, campsiteName: String) -> Unit,
+    onNavigateToWriteReview: (campsiteId: Long, campsiteName: String) -> Unit,
     onNavigateUp: () -> Unit,
     viewModel: ReservationListViewModel = hiltViewModel()
 ) {
